fix: latitude NaN
The userId is not anymore provided at the req body, but instead it's added to the req after a successfull authentication in loginRequired and i forgot to reflect this change here. Also, as the parseReqBody returns an array, latitude was getting the value of "userId" and "longitude" from "latitude" at the line 11 (considering the before scenario, where userId was at the req body).