refactor: User model snake to camelCase attribute
since in the sequelize's database config it's being used 'underscored: true', then there's no need to let any model attribute as snake_case by rule. But in this case 'selected_avatar' is snake-cased because the association alias and the foreign key name can't hold the same value, and making the user provide a snake-cased attribute when all the others are camel-cased is not a good idea. So i made a way for the model property to be snake-cased (selected_avatar), and request fields received in create and update routes to be used at the User model instance creation to be camel-cased (selectedAvatar).