fix: selected_avatar fk error
Fk constraint violations are are not thrown as validation errors, such as uk constraint violations, in sequelize. So a custom validation is now applied at the model level to handle uk violations before getting to the database, making the validation of both constraint be handled equally in the errorHandler middleware.