feat: birthDate column in users
In order to use in statistics about the application, like measuring the avarage age of users that are frequently interacting with the it, then the birthDate field is now required for all users. - all these changes in the commit are related to considering the birthDate as part of the user entity.