refactor: loginRequired add user directly in req
instead of attaching to the req object 'userId' and then using it to try to get the user from that id, now 'user' is passed directly, making this way simpler to show, update or detroy it. Also it's possible to create a model instance based on a model relationed to it as a parent model, so that's what is happening now in photo controller 'store' method: photos are created based on the authenticated user due to the relationship between the models.