feat: add csv routes for generate and download report about most active users
Routes 'process-csv/mostActiveUsers', 'download-csv/mostActiveUsers' are available to extract data from the database about the most active users - criteria based on the lastActivity, xp and photos amount the user has - and generate .csv files to be used in data analysis. The files generated are going to be located at src/reports folder. To achieve this, the library csv-stringify has been installed as a dependency.